+++ MapDriveInator/MapDriveInator.mm
@@ -23,7 +23,7 @@
 <attribute NAME="updateUrl" VALUE="${homepage}/releases/latest/download/version.properties"/>
 <attribute NAME="addonsMenu" VALUE="/menu_bar/edoTools"/>
 <attribute NAME="downloadUrl" VALUE="${homepage}/releases/download/${version}/"/>
-<attribute NAME="changelogUrl" VALUE=""/>
+<attribute NAME="changelogUrl" VALUE="${homepage}/releases/latest/download/history.md"/>
 <richcontent TYPE="NOTE" CONTENT-TYPE="xml/">
 <html>
   <head>
@@ -214,7 +214,7 @@
 <node TEXT="old script version packed as an Add-on" ID="ID_1092003668"/>
 </node>
 <node TEXT="v0.0.4" ID="ID_1109048795">
-<node TEXT="Added MDI Styles Template map" ID="ID_93006869"/>
+<node TEXT="Added Template map" ID="ID_93006869"/>
 <node TEXT="changed obtainBaseFolder logic" ID="ID_934053051"/>
 <node TEXT="corrected similarity issue" ID="ID_1643559503"/>
 <node TEXT="change folder_node recognition logic" ID="ID_345068445"/>